test: Analyzer 테스트 코드 AnalyzerTest로 분리
